Reputation tests  spendings progress
Spendings are  still not working exactly the same as in Java, but now the differences are not that big. The most major difference is that sometimes Java returns the ranks of both sellers and buyers and sometimes of just sellers. Python always returns only ranks of sellers. This is when spendigs>0. The ranks are the same in both implementations though. Given that calculations are the same, it does not seem to be a critical issue to me.